requirements.txt
kafka-python
psycopg2-binary
pandas
scikit-learn
flask
sqlalchemy
